Sortable Table exercise
first I added an event listener to the grid.
then I declared the function that goes in the event listener.

To sort the table, I declared a compareFunction based on what th heading was clicked. The compare functions differ whether I am comparing numbers or letters.

I covert the values of the table into an array so I can sort them. then I append the sorted array to the table.