Tool Tip exercise
I use the mouseover event over the whole document. if the target I'm hovering over does not have the tool tip attribute I return out of the function.

I created the tool tip by creating a element.

the element content is the content in the tool tip attribute.

then I set the positioning of the element.

I use the mouse out event to get rid of the element