Image Gallery exercise
I created an image gallery. I used delegation, meaning, in this case, if you click anywhere on the list it will fire an action. My safety statement was that if the target.event doesn't have a 'closest' 'a' element, return out of the function. Hence, no task is preformed by the function. Then I go on to prevent the browser default action of the 'a' element. Finally, I change the value of the src attribute of the element with the id largeImg to the href of the selected image. This ultimately changes the image being displayed.