Selectable List exercise
I completed the Selectable List exercise.

I used delegation over ul, and used the tagName to create a safety statement. I made it so if the ctrl or meta key was not pressed at the time of the click, all the previously selected classes would be removed. Then I just go on to add the class selected to the target.