Smart Tooltip exercise
I created it using mouseenter and mouseleave as apposed to mouseover and mouseout. This makes it so it ignores all the inner elements. the downside of this is you would not be able to use delegation. The tool tip only displays if you've hovered over the element a certain amount of time.